Safari-609.1.11.1/LayoutTests/accessibility/table-with-missing-aria-role-rows-expected.txt

a	b
a	b
This tests if a table uses the grid role, but forgets to add role='row' or role='gridcell' to native tr/td elements, it still exposes the right numbers of rows/cells.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS table.rowCount is 2
PASS table.columnCount is 2
PASS successfullyParsed is true

TEST COMPLETE
